@@ -22,24 +22,24 @@
     "url": "https://github.com/ROTBOW/the_shadow_realm/issues"
   },
   "dependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "axios": "^0.22.0",
-    "babel-loader": "^8.2.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
+    "@babel/core": "^7.22.10",
+    "@babel/preset-env": "^7.22.10",
+    "@babel/preset-react": "^7.22.5",
+    "axios": "^1.4.0",
+    "babel-loader": "^9.1.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.10.1",
+    "react-router": "^6.15.0",
+    "react-router-dom": "^6.15.0",
     "sass": "^1.66.1",
-    "webpack": "^5.58.1"
+    "webpack": "^5.88.2"
   },
   "devDependencies": {
-    "css-loader": "^6.4.0",
-    "http-server": "^13.0.2",
-    "sass-loader": "^12.6.0",
-    "style-loader": "^3.3.0",
-    "webpack-cli": "^4.9.0"
+    "css-loader": "^6.8.1",
+    "http-server": "^14.1.1",
+    "sass-loader": "^13.3.2",
+    "style-loader": "^3.3.3",
+    "webpack-cli": "^5.1.4"
   }
 }
